feat: inline external references during validation construction
This change is to fix a problem that occurs when downstream projects
attempt to validate openRPC documents. They need a way to resolve
in document references.

This change should have negligible performance impact, because the
resolution of the doucment is only performed on validation construction.
The cost of resolving the references, is required for the validation
process, so it should not add avoidable overhead.

Lastly this change should allow downstream components to resolve
references in a uniform way, without duplicating the process

fixes #125